requirements.txt
streamlit
numpy
opencv-python
tensorflow==2.17.0
pillow
